Project 2 - A file-persistent KV store
## This Commit

Writes Set and Remove commands to disk, performing naive compaction when
necessary. This allows for persistence in the case of a crash and more
stable performance over lots of commands.

## TODO

It seems like bitcask uses multiple files so I should surely do that
too. Also, my compaction is very naive. That'll be fun to optimize.