Use multiple log files to store commands
## This Commit

Updates `KvStore` to operate on a directory and write to multiple files.
One file is active and is written to. The other files are immutable and
only read from. Compaction only occurs on immutable files. This roughly
follows the design from [this `bitcask` intro dock][bitcask].

## Why?

This matches the original design more closely. I don't totally know the
pros and cons of this vs having a single file but I wonder if it has to
do with when there are multiple bitcask servers around or concurrent
read requests on a single server or something. We'll see!

[bitcask]: https://github.com/basho/bitcask/blob/develop/doc/bitcask-intro.pdf