MVP for configurable compaction policies
## This Commit

Adds some notes and skeleton structure for configuring a `KvStore`
instance with different compaction policies.

## Why?

Largely because I'm not sure how we should choose when to compact
immutable files. So this should make it maybe easier to write a harness
that tests different workloads. This will either show that different
workloads prefer different policies (in which case the policy
architecture can stay) or that one policy is the best for, generally
speaking, all workloads (in which case we can choose that policy with
confidence instead of guesswork).