Cache file length in memory
## This Commit

Attempts to cache file lengths in memory to avoid querying the
filesystem metadata on every write.

## Notes

Currently this adds back some String allocations I don't want. It also
emphasizes the need for `compactify` and `write_cmd` to share a similar
command-writing utility. Options here are:

1. Try to avoid the allocation because we know we're writing JSON and,
   given a particular `Command`, we know how long it'll be on disk
2. Try to use `Protocol` to `write` commands to the file. This is
   slightly wasted because `Protocol` doesn't leverage newlines and so
   has extra length information but sharing these might be nice because
   `Protocol::written_size` (or whatever) makes more sense than
   `Command::json_size`.

## Results

I didn't actually time anything which is shameful but `flamegraph` now
basically shows 100% of our application, in a write-heavy workload, is
receiving connections and updating the in-memory index so that's maybe
good.