Remove unnecessary `get` method on `KvStore`
## This Commit

Removes a duplicate implementation of the `get` method on the `KvStore`.

## Why?

I must have created it after the image of the standard library's
`HashMap` method which requires `Eq` and `Hash` on the key. Since we're
just going to use the string value of the key, however, we don't need
these requirements. In fact, the implementation of `KvsEngine` had the
correct implementation of `get` so we just remove the unneeded one.