Improve sharing of `protocol`
## This Commit

Updates `protocol::Cmd` in two ways:

1. updates the binary format to roughly match what's described in [the
   bitcask white paper][0]
2. adds a `read` method to allow reading from a reader in a hopefully
   efficient way, now that we have fixed key/value lengths

These changes should allow us to share `protocol::Cmd` with
`kvs::Command` (i.e., we should be able to use `protocol::Cmd` to send
commands over the network _and_ over the filesystem).

This commit also moves `protocol::Response` to `kvs-server`. I'm not
sure this is desirable, but it makes it more clear that this is a server
response mechanism, instead of a more general aspect of the protocol,
like `Cmd` is.

[0]: https://github.com/basho/bitcask/blob/develop/doc/bitcask-intro.pdf