Use `protocol::Cmd` in `kvs::KvStore`
## This Commit

Uses `protocol::Cmd` in place of `serde_json` for persisting commands
to, and reading from, disk.

## Why?

`cargo flamegraph` showed that most of our time in our read benchmark
was spent in `serde_json` reading data off disk. To optimize this, we
reuse `protocol::Cmd`. This seems to show a ~50% improvement from 15ms
to 7ms.