Benchmark writing and consolidate writes
## This Commit

Updates the `kvs-server` benchmark to do many more writes so we can see
that data more clearly. Once doing this, I saw lots of calls to
`write_all` and wondered if buffering those would help, so I tried
setting that up, though the approach is questionable at best.

## Results

That did nothing. Which makes sense. In hindsight, we were already
buffering so we could append the newline. Next, we're going to try
un-buffering and seeing if that magically fixes something.