Temporary mid-step: Extract ActiveFile struct
## This Commit

Extracts an `ActiveFile` struct for writing commands to.

## Why?

I think this is a good data hiding technique so that if we choose to
change the internals of `ActiveFile` to be a memory mapped file or an
internal buffer, the rest of the code Just Works. It's a little rough
now, and there's some copypasta, and there should probably be an
equivalent version for compacting/immutable files or something, but
let's start here, try the memory mapped file, see if it's an
improvement, and then decide what to do.