Clarify protocol crate
## This Commit

Updates the `protocol` crate so it's more clear what it's for by:

1. Improving documentation to explain the relationship to
   requests/responses to/from the database.
2. Improving documentation for implementation details of the protocols.
3. Adding more variants to the `Response` enum to identify a successful
   set from a successful get whose key is the empty string.
4. Extracting constants instead of magic byte literals.

## Why?

Every other day I opened this crate and forgot what it was for. I kept
thinking it was a strange mix of responsibilities. It turns out it
isn't, and I wanted to clearly document that.

Also, we'll want to update the