Remove MaxDeadRecordPolicy
## This Commit

Moves the idea of a max dead record policy into a comment.

## Why?

It wasn't used, wasn't good in its current state, and was adding
unnecessary noise to the code that needed to be worked around. For now,
let's get rid of it, and revisit compaction policies if it ever seems
worth it.