Clean up client binary with `Client` struct
## This Commit

Introduces a `Client` API in the library to be used by the `kvs-client`
binary.

## Why?

To clean up the binary and get more practice writing APIs.

This one is pretty iffy:

- Not exactly sure where the connecting should happen or what the return
  types should be.
- Also don't love the handling of the `Response` enum.

But the binary looks clean, so that's half the battle. We'll see if
there's ever a chance to test its ability to be extended gracefully.