Use `CmdReader` to optimize compaction
## This Commit

Very sloppily uses `CmdReader` to avoid UTF-8 validation when performing
compaction. The result appears to be a 55% reduction in the write
benchmark for `KvStore`.

## Why?

When performing compaction, we just need to copy the bytes from the
immutable file directly to the compacting file -- we don't need to
create a fully formed `String` object in memory and perform UTF-8
validation.